+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -1,0 +1,58 @@
+package com.techcourse.service;
+
+import com.interface21.dao.DataAccessException;
+import com.interface21.jdbc.datasource.DataSourceUtils;
+import com.interface21.transaction.support.TransactionSynchronizationManager;
+import com.techcourse.config.DataSourceConfig;
+import com.techcourse.domain.User;
+import java.sql.Connection;
+import java.sql.SQLException;
+import javax.sql.DataSource;
+
+public class TxUserService implements UserService {
+
+    private final UserService userService;
+    private final DataSource dataSource;
+
+    public TxUserService(final UserService userService) {
+        this.userService = userService;
+        this.dataSource = DataSourceConfig.getInstance();
+    }
+
+    @Override
+    public User findById(final long id) {
+        return userService.findById(id);
+    }
+
+    @Override
+    public void save(User user) {
+        userService.save(user);
+    }
+
+    @Override
+    public void changePassword(long id, String newPassword, String createBy) {
+        Connection connection = DataSourceUtils.getConnection(dataSource);
+        try {
+            connection.setAutoCommit(false);
+            TransactionSynchronizationManager.bindResource(dataSource, connection);
+            userService.changePassword(id, newPassword, createBy);
+            connection.commit();
+        } catch (Exception e) {
+            rollback(connection);
+            throw new DataAccessException("Transaction failed", e);
+        } finally {
+            TransactionSynchronizationManager.unbindResource(dataSource);
+            DataSourceUtils.releaseConnection(connection, dataSource);
+        }
+    }
+
+    private void rollback(Connection connection) {
+        if (connection != null) {
+            try {
+                connection.rollback();
+            } catch (SQLException e) {
+                throw new DataAccessException("Rollback failed", e);
+            }
+        }
+    }
+}
